\name{TOMproject-package}
\alias{TOMproject-package}
\alias{TOMproject}
\docType{package}
\title{A package for a simulation I'm doing.
}
\description{The simulation about Topological Overlap in personality psychology.
}
\details{
\tabular{ll}{
Package: \tab TOMproject\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-04\cr
License: \tab What license is it under?\cr
}
}
\author{
Giulio Costantini

Maintainer: Giulio Costantini <costantinigiulio@gmail.com>
}
\references{
}
\keyword{ Topological overlap }
\keyword{ Personality Psychology }